rest resource / brick client - delete method

+++ legoManager-rest/src/main/java/cz/muni/fi/PA165/rest/BrickResource.java
@@ -1,19 +1,13 @@
 package cz.muni.fi.PA165.rest;
 
 import cz.muni.fi.PA165.api.dto.BrickDto;
+import cz.muni.fi.PA165.api.dto.BuildingKitDto;
 import cz.muni.fi.PA165.api.service.BrickService;
+import cz.muni.fi.PA165.api.service.BuildingKitService;
 import cz.muni.fi.PA165.api.service.Color;
 import org.springframework.beans.factory.annotation.Autowired;
 
-import javax.ws.rs.Consumes;
-import javax.ws.rs.DELETE;
-import javax.ws.rs.GET;
-import javax.ws.rs.POST;
-import javax.ws.rs.PUT;
-import javax.ws.rs.Path;
-import javax.ws.rs.PathParam;
-import javax.ws.rs.Produces;
-import javax.ws.rs.QueryParam;
+import javax.ws.rs.*;
 import javax.ws.rs.core.GenericEntity;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
@@ -28,6 +22,8 @@ public class BrickResource {
 
     @Autowired
     BrickService brickService;
+    @Autowired
+    BuildingKitService buildingKitService;
 
     @POST
     @Consumes(MediaType.APPLICATION_JSON)
@@ -61,7 +57,8 @@ public class BrickResource {
         }
 
         if (brickDtoList != null) {
-            GenericEntity<List<BrickDto>> brickEntity = new GenericEntity<List<BrickDto>>(brickDtoList) {};
+            GenericEntity<List<BrickDto>> brickEntity = new GenericEntity<List<BrickDto>>(brickDtoList) {
+            };
 
             return Response.status(Response.Status.OK).entity(brickEntity).build();
         } else {
@@ -70,20 +67,51 @@ public class BrickResource {
     }
 
 
+    /**
+     * delete brick
+     *
+     * @param id Long id of brick
+     * @return String confirmation of deleted brick or it cant be deleted because
+     * brick is used by some of building kits, returns names of this building kits
+     */
     @DELETE
-    @Produces(MediaType.APPLICATION_JSON)
     @Path("{id}")
+    @Consumes(MediaType.APPLICATION_JSON)
+    @Produces(MediaType.APPLICATION_JSON)
     public Response delete(@PathParam("id") Long id) {
-        brickService.delete(id);
-        return Response.status(Response.Status.OK).build();
+
+        BrickDto brick = brickService.findById(id);
+
+        // if brick with this id does not exit, return s
+        //TODO problem na servisni vrstve
+        if(brick == null) {
+            return Response.status(Response.Status.NO_CONTENT).build();
+        }
+
+        // find out if brick is used by some building kit first
+        // otherwise removal of used brick would violate the constraint
+        List<BuildingKitDto> buildingKitDtoList = buildingKitService.findByBrick(brick);
+
+        // list is empty, brick is not contained in any building kit => delete brick
+        if (buildingKitDtoList.isEmpty()) {
+            brickService.delete(id);
+            return Response.status(Response.Status.OK).build();
+        }
+
+        // list is not empty == brick is used by some building kit
+        // list that building kits
+        GenericEntity<List<BuildingKitDto>> genericEntityList = new GenericEntity<List<BuildingKitDto>>(buildingKitDtoList) {
+        };
+        return Response.status(Response.Status.CONFLICT).entity(genericEntityList).build();
+
     }
 
     @GET
     @Path("{id}")
     @Produces(MediaType.APPLICATION_JSON)
     public Response findById(@PathParam("id") Long id) {
-        BrickDto brickdDto = brickService.findById(id);
-        return Response.status(Response.Status.OK).entity(brickdDto).build();
+        BrickDto brickDto = brickService.findById(id);
+        return Response.status(Response.Status.OK).entity(brickDto).build();
     }
 
 }
